stock-notifcation/requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
fonttools==4.57.0
idna==3.10
importlib_metadata==8.7.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
squarify==0.4.4
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vnai==2.0.4
vnstock==3.2.4
vnstock3==3.2.1
vnstock_ezchart==0.0.2
wordcloud==1.9.4
zipp==3.21.0
